requirements.txt
streamlit 
streamlit-chat
groq
python-dotenv
asyncio
